新增文件操作接口，补充Vector3上限 API

+++ include/d3d12helper/Utils.h
@@ -41,8 +41,10 @@ inline void SetDXGIDebug(IDXGIObject *object, std::string name) {
   object->SetPrivateData(WKPDID_D3DDebugObjectName, name.size(), name.data());
 }
 
-inline uint32_t UpperMemorySize(uint32_t size, uint32_t alignment) {
+inline uint64_t UpperMemorySize(uint64_t size, uint64_t alignment) {
   return (size + alignment - 1) & ~(alignment - 1);
 }
+
+
 
 } // namespace d3dhelper
